+++ libs/components/src/lib/components/bracket/components/bracket/bracket.component.types.ts
@@ -1,0 +1,6 @@
+import { BracketMatch } from '../../types';
+
+export interface ConnectedMatches {
+  previousMatches: (BracketMatch | null)[] | null;
+  nextMatch: BracketMatch | null;
+}