+++ libs/components/src/lib/components/scrollable/components/scrollable/scrollable.component.ts
@@ -4,23 +4,27 @@ import {
   ChangeDetectionStrategy,
   Component,
   ElementRef,
+  EventEmitter,
   HostBinding,
   inject,
   Input,
   OnInit,
+  Output,
   Renderer2,
   ViewChild,
   ViewEncapsulation,
 } from '@angular/core';
-import { LetDirective, NgClassType } from '@ethlete/core';
+import {
+  CursorDragScrollDirective,
+  LetDirective,
+  NgClassType,
+  ObserveContentDirective,
+  ObserveScrollStateDirective,
+  ScrollObserverScrollState,
+} from '@ethlete/core';
 import { DestroyService } from '../../../../services';
 import { BehaviorSubject, takeUntil, tap } from 'rxjs';
-import {
-  ChevronIconComponent,
-  CursorDragScrollDirective,
-  ObserveScrollStateDirective,
-  ScrollableScrollState,
-} from '../../partials';
+import { ChevronIconComponent } from '../../../icons';
 
 @Component({
   selector: 'et-scrollable',
@@ -29,7 +33,15 @@ import {
   standalone: true,
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [CursorDragScrollDirective, ObserveScrollStateDirective, NgClass, NgIf, LetDirective, ChevronIconComponent],
+  imports: [
+    CursorDragScrollDirective,
+    ObserveScrollStateDirective,
+    NgClass,
+    NgIf,
+    LetDirective,
+    ChevronIconComponent,
+    ObserveContentDirective,
+  ],
   host: {
     class: 'et-scrollable',
   },
@@ -81,10 +93,13 @@ export class ScrollableComponent implements OnInit {
   }
   private _renderScrollbars = false;
 
+  @Output()
+  readonly contentChanged = new EventEmitter<MutationRecord[]>();
+
   @ViewChild('scrollable', { static: true })
   scrollable!: ElementRef<HTMLElement>;
 
-  protected readonly scrollState$ = new BehaviorSubject<ScrollableScrollState | null>(null);
+  protected readonly scrollState$ = new BehaviorSubject<ScrollObserverScrollState | null>(null);
 
   ngOnInit(): void {
     this.scrollState$
@@ -105,8 +120,12 @@ export class ScrollableComponent implements OnInit {
       .subscribe();
   }
 
-  protected _scrollStateChanged(scrollState: ScrollableScrollState) {
+  protected _scrollStateChanged(scrollState: ScrollObserverScrollState) {
     this.scrollState$.next(scrollState);
+  }
+
+  protected _contentChanged(data: MutationRecord[]) {
+    this.contentChanged.emit(data);
   }
 
   protected scrollOneContainerSizeToStart() {