+++ libs/components/src/lib/components/scrollable/partials/observe-scroll-state/observe-scroll-state.types.ts
@@ -1,7 +1,0 @@
-export interface ScrollableScrollState {
-  isAtStart: boolean;
-  isAtEnd: boolean;
-  canScroll: boolean;
-}
-
-export type ObservedScrollableChild = 'first' | 'last';