+++ libs/components/src/lib/components/scrollable/public-api.ts
@@ -1,5 +1,3 @@
 export * from './components/public-api';
-export * from './partials/public-api';
-export * from './utils/public-api';
 
 export * from './scrollable.module';