+++ libs/components/src/lib/components/tabs/public-api.ts
@@ -1,0 +1,7 @@
+export * from './animations/public-api';
+export * from './components/public-api';
+export * from './partials/public-api';
+export * from './types/public-api';
+export * from './utils/public-api';
+
+export * from './tabs.module';