+++ libs/components/src/lib/components/tabs/types/public-api.ts
@@ -1,0 +1,1 @@
+export * from './tab-config.types';